+++ src/collective/iconifiedcategory/utils.py
@@ -391,18 +391,6 @@ def print_message(obj=None, to_print_value=None):
         return messages[to_print_value]
 
 
-def confidential_message(obj=None, confidential_value=None):
-    """Return the confidential status message for the given object"""
-    messages = {
-        True: u'Element is confidential',
-        False: u'Element is not confidential',
-    }
-    if obj:
-        return messages.get(getattr(obj, 'confidential', None), '')
-    else:
-        return messages[confidential_value]
-
-
 def signed_message(obj=None, to_sign_value=None, signed_value=None):
     """Return the signed message for the given object"""
     messages = {
@@ -420,16 +408,16 @@ def signed_message(obj=None, to_sign_value=None, signed_value=None):
         return messages[signed_value]
 
 
-def publishable_message(obj=None, publishable_value=None):
-    """Return the publishable status message for the given object"""
+def boolean_message(obj=None, attr_name='', value=None):
+    """Return the default boolean status message for the given object"""
     messages = {
-        True: u'Element is publishable',
-        False: u'Element is not publishable',
+        True: u'Element is {0}'.format(attr_name),
+        False: u'Element is not {0}'.format(attr_name),
     }
     if obj:
-        return messages.get(getattr(obj, 'publishable', None), '')
+        return messages.get(getattr(obj, attr_name, None), '')
     else:
-        return messages[publishable_value]
+        return messages[value]
 
 
 @ram.cache(lambda f, p: (p, time() // (60 * 60)))